feature: add custom avatar ability in search
